Aula 2 -Avançando no tratamento de erros

+++ src/controllers/autorController.js
@@ -1,62 +1,74 @@
+import ErroNaoEncontrado from "../erros/ErroNaoEncontrado.js";
 import { autor } from "../models/Autor.js";
-import mongoose from "mongoose";
+
 
 class AutorController {
 
-  static async listarAutores (req, res) {
+  static async listarAutores (req, res, next) {
     try {
       const listaAutores = await autor.find({});
       res.status(200).json(listaAutores);            
     } catch (erro) {
-      res.status(500).json({ message: `${erro.message} - falha na requisição` });
+      next(erro);
     }
   }
 
-  static async listarAutorPorId (req, res) {
+  static async listarAutorPorId (req, res, next) {
     try {
       const id = req.params.id;
       const autorEncontrado = await autor.findById(id);
 
       if (autorEncontrado !== null) {
         res.status(200).json(autorEncontrado);            
-      } else {
-        res.status(404).send({message: "Id do Autor não localizado."});
+      } 
+      else {
+        next(new ErroNaoEncontrado("Id do Autor não localizado."));
       }
     } catch (erro) {
-      if (erro instanceof mongoose.Error.CastError) {
-        res.status(400).json({ message: "um ou mais dados fornecidos estão incorretos." });  
-      }else {
-        res.status(500).json({ message: `${erro.message} - falha na requisição` });
-      }
+      next(erro);
     }
   }
 
-  static async cadastrarAutor (req, res) {
+  static async cadastrarAutor (req, res, next) {
     try {
       const novoAutor = await autor.create(req.body);   
-      res.status(201).json({ message: "criado com sucesso", autor: novoAutor });
+      res.status(201).json({ message: "Criado com sucesso", autor: novoAutor });
     } catch (erro) {
-      res.status(500).json({ message: `${erro.message} - falha ao castrar Autor` });
+      next(erro);
     }
   }
 
-  static async atualizarAutor (req, res) {
+  static async atualizarAutor (req, res, next) {
     try {
       const id = req.params.id;
-      await autor.findByIdAndUpdate(id, req.body);
-      res.status(200).json({message: "autor atualizado"});            
+      const autorEncontrado = await autor.findById(id);
+
+      if (autorEncontrado !== null) {
+        await autor.findByIdAndUpdate(id, req.body);
+        res.status(200).json({message: "Autor atualizado", status: 200}); 
+      } 
+      else {
+        next(new ErroNaoEncontrado("Id do Autor não localizado."));
+      }
     } catch (erro) {
-      res.status(500).json({ message: `${erro.message} - falha na requisição` });
+      next(erro);
     }
   }
 
-  static async excluirAutor (req, res) {
+  static async excluirAutor (req, res, next) {
     try {
       const id = req.params.id;
-      await autor.findByIdAndDelete(id);
-      res.status(200).json({message: "autor excluido com sucesso"});            
+      const autorEncontrado = await autor.findById(id);
+
+      if (autorEncontrado !== null) {
+        await autor.findByIdAndDelete(id);
+        res.status(200).json({message: "Autor excluido com sucesso", status: 200}); 
+      } 
+      else {
+        next(new ErroNaoEncontrado("Id do Autor não localizado."));
+      }
     } catch (erro) {
-      res.status(500).json({ message: `${erro.message} - falha na exclusão` });
+      next(erro);
     }
   }
 }
